refactor: Update TestsPage component to include FaGripVertical icon in SidebarGrids
This commit updates the TestsPage component to include the FaGripVertical icon from the react-icons/fa library in the SidebarGrids component. The icon is added to each grid item to indicate that it can be dragged and rearranged. This change enhances the visual representation and usability of the SidebarGrids in the TestsPage.