refactor: Improve TestsPage component by making children span element a flex container
This commit improves the TestsPage component by making the children span element inside the DraggableButton component a flex container. By adding the "flex" class to the span element, we enable flexbox layout for the children, allowing them to be properly aligned and spaced. This change enhances the visual presentation and flexibility of the DraggableButton component.